+++ ts/libs/polaris-nx/src/executors/docker-build/schema.d.ts
@@ -1,0 +1,3 @@
+export interface GenerateDockerBuildExecutorSchema {
+    projectName: string;
+}